+++ crates/core/Cargo.toml
@@ -111,8 +111,8 @@ enum-map-derive = "0.17.0"
 gethostname = "0.4.3"
 humantime = "2.1.0"
 itertools = "0.12.0"
-zstd = "0.13.0"
 quick_cache = "0.4.0"
+zstd = "0.13.0"
 
 [target.'cfg(not(windows))'.dependencies]
 sha2 = { version = "0.10", features = ["asm"] }
