+++ crates/core/src/commands.rs
@@ -10,7 +10,6 @@ pub mod copy;
 /// The `dump` command.
 pub mod dump;
 pub mod forget;
-pub mod fs;
 pub mod init;
 pub mod key;
 pub mod merge;
